+++ photon-model-tasks/src/test/java/com/vmware/photon/controller/model/tasks/ResourceEnumerationTaskServiceTest.java
@@ -138,7 +138,6 @@ public class ResourceEnumerationTaskServiceTest extends Suite {
         public void testServiceOptions() {
 
             EnumSet<Service.ServiceOption> expected = EnumSet.of(
-                    Service.ServiceOption.IDEMPOTENT_POST,
                     Service.ServiceOption.INSTRUMENTATION,
                     Service.ServiceOption.OWNER_SELECTION,
                     Service.ServiceOption.REPLICATION);